Fix up -compat symlinks for removal of aspect ratio-specific images
For now, at least, keep all the same filenames even though they
all point to the same image and the ratios are a lie. Not sure
if it's safe to drop them all. We could do some fancy conversion
or something but I don't want to do anything too complex without
the design team looking at it.

Signed-off-by: Adam Williamson <awilliam@redhat.com>